refactor: Character DTO 패키지
