+++ fourtooncookie/src/main/java/com/startingblue/fourtooncookie/character/CharacterService.java
@@ -5,8 +5,8 @@ import com.startingblue.fourtooncookie.artwork.domain.Artwork;
 import com.startingblue.fourtooncookie.character.domain.Character;
 import com.startingblue.fourtooncookie.character.domain.CharacterVisionType;
 import com.startingblue.fourtooncookie.global.domain.PaymentType;
-import com.startingblue.fourtooncookie.character.dto.request.CharacterSaveRequest;
-import com.startingblue.fourtooncookie.character.dto.request.CharacterUpdateRequest;
+import com.startingblue.fourtooncookie.character.dto.CharacterSaveRequest;
+import com.startingblue.fourtooncookie.character.dto.CharacterUpdateRequest;
 import com.startingblue.fourtooncookie.character.exception.CharacterDuplicateException;
 import com.startingblue.fourtooncookie.character.exception.CharacterNotFoundException;
 import lombok.RequiredArgsConstructor;
