+++ fourtooncookie/src/main/java/com/startingblue/fourtooncookie/character/dto/CharacterSavedResponses.java
@@ -1,4 +1,4 @@
-package com.startingblue.fourtooncookie.character.dto.response;
+package com.startingblue.fourtooncookie.character.dto;
 
 import com.startingblue.fourtooncookie.character.domain.Character;
