+++ fourtooncookie/src/test/java/com/startingblue/fourtooncookie/character/service/CharacterServiceTest.java
@@ -7,10 +7,10 @@ import com.startingblue.fourtooncookie.character.domain.Character;
 import com.startingblue.fourtooncookie.character.CharacterRepository;
 import com.startingblue.fourtooncookie.character.domain.CharacterVisionType;
 import com.startingblue.fourtooncookie.global.domain.PaymentType;
-import com.startingblue.fourtooncookie.character.dto.request.CharacterSaveRequest;
-import com.startingblue.fourtooncookie.character.dto.request.CharacterUpdateRequest;
-import com.startingblue.fourtooncookie.character.dto.response.CharacterSavedResponse;
-import com.startingblue.fourtooncookie.character.dto.response.CharacterSavedResponses;
+import com.startingblue.fourtooncookie.character.dto.CharacterSaveRequest;
+import com.startingblue.fourtooncookie.character.dto.CharacterUpdateRequest;
+import com.startingblue.fourtooncookie.character.dto.CharacterSavedResponse;
+import com.startingblue.fourtooncookie.character.dto.CharacterSavedResponses;
 import com.startingblue.fourtooncookie.character.exception.CharacterNotFoundException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
